+++ charitable/includes/admin/views/settings/password.php
@@ -24,5 +24,5 @@ endif;
 	class="<?php echo esc_attr( $view_args['classes'] ); ?>"
 	<?php echo esc_attr( charitable_get_arbitrary_attributes( $view_args ) ); ?> />
 <?php if ( isset( $view_args['help'] ) ) : ?>
-	<div class="charitable-help"><?php echo esc_html( $view_args['help'] ); ?></div>
+	<div class="charitable-help"><?php echo ( $view_args['help'] ); ?></div>
 <?php endif; ?>